Move common classes from sample to commons
