+++ zowe-rest-api-commons-spring/src/main/java/org/zowe/commons/apidoc/ApiDocConstants.java
@@ -7,12 +7,12 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.sample.apiservice.zowe;
+package org.zowe.commons.apidoc;
 
-import com.ca.mfaas.eurekaservice.client.config.ApiMediationServiceConfig;
+public final class ApiDocConstants {
+    public static final String DOC_SCHEME_BASIC_AUTH = "basicAuth";
 
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("apiml.service")
-public class ApiMediationServiceConfigBean extends ApiMediationServiceConfig {
+    private ApiDocConstants() {
+        // Utility class
+    }
 }
