+++ zowe-rest-api-commons-spring/src/main/java/org/zowe/commons/spring/security/CsrfController.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.sample.apiservice;
+package org.zowe.commons.spring.security;
 
 import org.springframework.security.web.csrf.CsrfToken;
 import org.springframework.web.bind.annotation.GetMapping;
