+++ zowe-rest-api-commons-spring/src/test/java/org/zowe/commons/apiml/RegisterToApiLayerTests.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package org.zowe.sample.apiservice.zowe;
+package org.zowe.commons.apiml;
 
 import java.security.InvalidParameterException;
